requirements.txt
fastapi==0.110.0
joblib==1.3.2
numpy==1.25.2
protobuf==5.26.0
pydantic==2.6.4
scikit_learn==1.3.2
uvicorn==0.28.0
google-cloud-storage==2.5.0
google-auth==2.12.0
load-dotenv==0.1.0